Conflicts:
	js/src/asmjs/AsmJSFrameIterator.cpp
	js/src/asmjs/AsmJSValidate.cpp
	js/src/irregexp/NativeRegExpMacroAssembler.cpp
	js/src/irregexp/NativeRegExpMacroAssembler.h
	js/src/jit/BaselineCompiler.cpp
	js/src/jit/BaselineDebugModeOSR.cpp
	js/src/jit/BaselineIC.cpp
	js/src/jit/CodeGenerator.cpp
	js/src/jit/IonCaches.cpp
	js/src/jit/MacroAssembler.cpp
	js/src/jit/MacroAssembler.h
	js/src/jit/Registers.h
	js/src/jit/shared/CodeGenerator-shared.cpp
	js/src/jit/x86-shared/Assembler-x86-shared.h

+++ js/src/jit/arm64/Architecture-arm64.cpp
@@ -28,7 +28,7 @@ Registers::FromName(const char *name)
     if (strcmp(name, "x31") == 0) // Default name "sp"
         return sp;
 
-    for (size_t i = 0; i < Total; i++) {
+    for (uint32_t i = 0; i < Total; i++) {
         if (strcmp(GetName(i), name) == 0)
             return Code(i);
     }
